Update defines.h
Adjusted TMP_RPM_CAP to 9000rpm. This makes the adjustable RPM using the potentiometer the same as the GUI interface. Also found when maxing out the physical potiemeter knob for RPM that the unit would stay at max RPM value and not return down with the knob. Adjusting the cap to match the GUI fixed that bug.