Update README.md
Updated the README to let people know (hopefully properly) that this is a fork of the Speeduino Ardu-stim of Josh Stewarts, which is also a fork of the original ardu-stim. Also stated I've been adding wheel patterns, and letting folks know to please copy/paste or use my fork of this as they see fit. Also added my twisted builds logo